requirements.txt
Django>=5.0
gunicorn
torch
torchvision
Pillow
